Add BaseMCRTensor
Add a new base class to support future implementation of variations of
MCR proposals.
